// src/dal/enums/tg/district-type.ts
export enum DistrictType {
    NONE = 0,
    VOENKA = 1,
    VOSTOCHNIY = 2,
    KINDIYKA = 4,
    STEKLOTARA = 8,
    TEKSTILNIY = 16,
    HBK = 32,
    ZHILMASIV = 64,
    ZHILPOSELOK = 128,
    ZABALKA = 256,
    NEFTEGAVAN = 512,
    OSTROV = 1024,
    CENTR = 2048,
    PORT_ELEVATOR = 4096,
    PRIVOKZALNIY = 8192,
    SHUMENSKIY = 16384,
    MELNIZA = 32768,
    SEVERNIY_TAVRICHESK = 65536,
}